refactor: update Dispatcher imports to use core.network package for consistency
