+++ core/data/src/main/java/com/keelim/core/data/source/firebase/FirebaseRepositoryImpl.kt
@@ -6,8 +6,8 @@ import com.google.firebase.FirebaseApp
 import com.google.firebase.database.Logger
 import com.google.firebase.database.database
 import com.google.firebase.messaging.messaging
-import com.keelim.common.Dispatcher
-import com.keelim.common.KeelimDispatchers
+import com.keelim.core.network.Dispatcher
+import com.keelim.core.network.KeelimDispatchers
 import com.keelim.core.data.BuildConfig
 import com.keelim.data.repository.FirebaseRepository
 import com.keelim.model.EcoCalEntry
