+++ core/data/src/main/java/com/keelim/core/data/source/note/NoteRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.keelim.core.data.source.note
 
-import com.keelim.common.Dispatcher
-import com.keelim.common.KeelimDispatchers
+import com.keelim.core.network.Dispatcher
+import com.keelim.core.network.KeelimDispatchers
 import com.keelim.core.database.dao.NoteDao
 import com.keelim.core.database.mapper.toNotices
 import com.keelim.core.database.mapper.toNoticesEntity
