+++ core/data/src/main/java/com/keelim/core/data/source/notification/NotificationRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.keelim.core.data.source.notification
 
-import com.keelim.common.Dispatcher
-import com.keelim.common.KeelimDispatchers
+import com.keelim.core.network.Dispatcher
+import com.keelim.core.network.KeelimDispatchers
 import com.keelim.core.data.BuildConfig
 import com.keelim.core.network.di.KtorNetworkModule
 import com.keelim.core.network.model.NoticeResponse
