+++ core/network/src/main/java/com/keelim/core/network/KeelimDispatchers.kt
@@ -1,4 +1,4 @@
-package com.keelim.common
+package com.keelim.core.network
 
 import javax.inject.Qualifier
